fix 01

+++ Harl.cpp
@@ -22,9 +22,9 @@ void Harl::_error(void)
 
 void Harl::complain( std::string level )
 {
-    void (Harl::*f[4])(void) = { &Harl::_debug, &Harl::_info, &Harl::_warning, &Harl::_error };
-    std::string levels[4] = { "DEBUG", "INFO", "WARNING", "ERROR" };
-    for (int i = 0; i < 4; i++)
+    void(Harl::*f[4])(void) = {&Harl::_debug, &Harl::_info, &Harl::_warning ,&Harl::_error};
+    std::string levels[4] = {"DEBUG", "INFO", "WARNING", "ERROR"};
+    for (short i = 0; i < 4; i++)
     {
         if (levels[i] == level)
             (this->*f[i])();
